+++ calofficials/src/main/java/api/model/Senator.java
@@ -2,7 +2,8 @@ package api.model;
 
 /**
  * Senator class that extends Official.
- * April 19, 2026
+ * @author Kelvin Myat
+ * @since April 19, 2026
  */
 public class Senator extends Official {
 
@@ -16,7 +17,7 @@ public class Senator extends Official {
      * @param districtId the district id
      */
     public Senator(String name, String party, String email, String office,
-                   double salary, int districtId) {
+                   Long salary, Long districtId) {
         super(name, party, email, office, salary, districtId, "Senator");
     }
 }
